@@ -12,15 +12,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
+    "@angular/animations": "~14.3.0",
     "@angular/cdk": "~13.3.0",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
+    "@angular/common": "~14.3.0",
+    "@angular/compiler": "~14.3.0",
+    "@angular/core": "~14.3.0",
+    "@angular/forms": "~14.3.0",
+    "@angular/platform-browser": "~14.3.0",
+    "@angular/platform-browser-dynamic": "~14.3.0",
+    "@angular/router": "~14.3.0",
     "@crawl/canvas-helper": "0.2.2",
     "@datorama/akita": "7.1.1",
     "@fullcalendar/angular": "5.11.0",
@@ -71,7 +71,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "~13.3.5",
     "@angular/cli": "^13.3.5",
-    "@angular/compiler-cli": "~13.3.0",
+    "@angular/compiler-cli": "~14.3.0",
     "@ngneat/transloco-keys-manager": "3.4.1",
     "@types/jasmine": "~3.10.0",
     "@types/lodash": "^4.14.202",
