@@ -10,7 +10,7 @@ import {
   Component,
   OnInit
 } from '@angular/core';
-import {FormGroup, Validators, FormBuilder,} from '@angular/forms';
+import {UntypedFormGroup, Validators, UntypedFormBuilder,} from '@angular/forms';
 import {ModalState} from '@app/enums/modal-state.enum';
 import {
   AdminUsersService,
@@ -67,7 +67,7 @@ export class NewGroupModalComponent implements OnInit {
   public constructor(
     public readonly modalRef: DialogRef,
     public readonly notesService: NotesService,
-    private readonly fb: FormBuilder,
+    private readonly fb: UntypedFormBuilder,
     private readonly cdr: ChangeDetectorRef,
     private readonly translocoService: TranslocoService,
     private readonly toastrService: ToastrService,
@@ -76,7 +76,7 @@ export class NewGroupModalComponent implements OnInit {
   ) {
   }
 
-  private get f(): FormGroup['controls'] {
+  private get f(): UntypedFormGroup['controls'] {
     return this.form.controls;
   }
 
@@ -185,7 +185,7 @@ export class NewGroupModalComponent implements OnInit {
   }
 
   MustMatch(controlName: string, matchingControlName: string) {
-    return (formGroup: FormGroup) => {
+    return (formGroup: UntypedFormGroup) => {
       const control = formGroup.controls[controlName];
       const matchingControl = formGroup.controls[matchingControlName];
 
